Split the test_parse.py by language
- switch over to utf-8 encoding for all the files
