+++ test/unittests/util/test_parse_es.py
@@ -1,0 +1,81 @@
+# -*- coding: utf-8 -*-
+#
+# Copyright 2017 Mycroft AI Inc.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#    http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+import unittest
+
+from mycroft.util.parse import normalize
+
+
+class TestNormalize(unittest.TestCase):
+    """
+        Test cases for Spanish parsing
+    """
+    def test_articles_es(self):
+        self.assertEqual(normalize("esta es la prueba", lang="es",
+                                   remove_articles=True),
+                         "esta es prueba")
+        self.assertEqual(normalize("y otra prueba", lang="es",
+                                   remove_articles=True),
+                         "y otra prueba")
+
+    def test_numbers_es(self):
+        self.assertEqual(normalize("esto es un uno una", lang="es"),
+                         "esto es 1 1 1")
+        self.assertEqual(normalize("esto es dos tres prueba", lang="es"),
+                         "esto es 2 3 prueba")
+        self.assertEqual(normalize("esto es cuatro cinco seis prueba",
+                                   lang="es"),
+                         "esto es 4 5 6 prueba")
+        self.assertEqual(normalize(u"siete mï¿½s ocho mï¿½s nueve", lang="es"),
+                         u"7 mï¿½s 8 mï¿½s 9")
+        self.assertEqual(normalize("diez once doce trece catorce quince",
+                                   lang="es"),
+                         "10 11 12 13 14 15")
+        self.assertEqual(normalize(u"dieciséis diecisiete", lang="es"),
+                         "16 17")
+        self.assertEqual(normalize(u"dieciocho diecinueve", lang="es"),
+                         "18 19")
+        self.assertEqual(normalize(u"veinte treinta cuarenta", lang="es"),
+                         "20 30 40")
+        self.assertEqual(normalize(u"treinta y dos caballos", lang="es"),
+                         "32 caballos")
+        self.assertEqual(normalize(u"cien caballos", lang="es"),
+                         "100 caballos")
+        self.assertEqual(normalize(u"ciento once caballos", lang="es"),
+                         "111 caballos")
+        self.assertEqual(normalize(u"habï¿½a cuatrocientas una vacas",
+                                   lang="es"),
+                         u"habï¿½a 401 vacas")
+        self.assertEqual(normalize(u"dos mil", lang="es"),
+                         "2000")
+        self.assertEqual(normalize(u"dos mil trescientas cuarenta y cinco",
+                                   lang="es"),
+                         "2345")
+        self.assertEqual(normalize(
+            u"ciento veintitrés mil cuatrocientas cincuenta y seis",
+            lang="es"),
+            "123456")
+        self.assertEqual(normalize(
+            u"quinientas veinticinco mil", lang="es"),
+            "525000")
+        self.assertEqual(normalize(
+            u"novecientos noventa y nueve mil novecientos noventa y nueve",
+            lang="es"),
+            "999999")
+
+
+if __name__ == "__main__":
+    unittest.main()
